@@ -1,17 +1,17 @@
 name: flutter_widget_from_html_core
-version: 0.8.5+1
+version: 0.9.0+1
 description: Flutter package to render html as widgets that focuses on correctness and extensibility.
 homepage: https://github.com/daohoangson/flutter_widget_from_html/tree/master/packages/core
 
 environment:
-  flutter: ">=2.2.0"
+  flutter: ">=3.3.0"
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   csslib: ^0.17.0
   flutter:
     sdk: flutter
-  fwfh_text_style: ^2.2.0
+  fwfh_text_style: ^2.22.08+1
   html: ^0.15.0
 
 dependency_overrides:
