@@ -1,17 +1,17 @@
 name: fwfh_just_audio
-version: 0.6.2+2
+version: 0.9.0
 description: WidgetFactory extension to render AUDIO with the just_audio plugin.
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=2.0.0"
+  flutter: ">=2.2.0"
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_widget_from_html_core: ">=0.6.0 <0.9.0"
-  just_audio: ">=0.7.0 <0.10.0"
+  flutter_widget_from_html_core: ">=0.8.0 <0.10.0"
+  just_audio: ">=0.9.0 <0.10.0"
 
 dependency_overrides:
   flutter_widget_from_html_core:
@@ -22,6 +22,6 @@ dependency_overrides:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  just_audio_platform_interface: any
+  just_audio_platform_interface: ">=4.0.0 <5.0.0"
   lint: any
   tuple: any
