@@ -1,10 +1,10 @@
 name: fwfh_text_style
-version: 2.7.3+1
+version: 2.22.08+1
 description: A TextStyle replacement that allows resetting height behavior to normal.
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=2.6.0-0"
+  flutter: ">=3.1.0-0"
   sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
