@@ -1,5 +1,6 @@
 import 'package:demo_app/screens/hello_world.dart' as enhanced;
 import 'package:demo_app/widgets/popup_menu.dart';
+import 'package:demo_app/widgets/selection_area.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
 
@@ -7,12 +8,13 @@ class HelloWorldCoreScreen extends StatelessWidget {
   const HelloWorldCoreScreen({Key key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) => Scaffold(
+  Widget build(BuildContext context) => SelectionAreaScaffold(
         appBar: AppBar(
           title: const Text('HelloWorldCoreScreen'),
           actions: const [
             PopupMenu(
               scrollToTop: true,
+              toggleIsSelectable: true,
             ),
           ],
         ),
