@@ -373,8 +373,10 @@ Future<void> pumpWidget(WidgetTester tester, Widget child) async {
   tester.binding.window.devicePixelRatioTestValue = 1.0;
   addTearDown(tester.binding.window.clearDevicePixelRatioTestValue);
 
-  tester.binding.window.textScaleFactorTestValue = 1.0;
-  addTearDown(tester.binding.window.clearTextScaleFactorTestValue);
+  tester.binding.window.platformDispatcher.textScaleFactorTestValue = 1.0;
+  addTearDown(
+    tester.binding.window.platformDispatcher.clearTextScaleFactorTestValue,
+  );
 
   await tester.pumpWidget(MaterialApp(home: child));
   await tester.pump();
@@ -420,11 +422,9 @@ class _ListViewTestApp extends StatelessWidget {
 }
 
 class _SliverListTestApp extends StatelessWidget {
-  final String? html;
   final Key? keyBottom;
 
-  const _SliverListTestApp({this.html, Key? key, this.keyBottom})
-      : super(key: key);
+  const _SliverListTestApp({Key? key, this.keyBottom}) : super(key: key);
 
   @override
   Widget build(BuildContext _) => Scaffold(
@@ -432,7 +432,7 @@ class _SliverListTestApp extends StatelessWidget {
           cacheExtent: 0,
           slivers: [
             HtmlWidget(
-              html ?? htmlDefault,
+              htmlDefault,
               factoryBuilder: () => _WidgetFactory(),
               key: globalKey,
               renderMode: RenderMode.sliverList,
