@@ -18,7 +18,7 @@ void main() {
 
   group('renders progress indicator', () {
     const html = '<video><source src="$src"></video>';
-    Future<String> _explain(WidgetTester tester) async {
+    Future<String> progressIndicatorExplain(WidgetTester tester) async {
       final explained = await explain(
         tester,
         html,
@@ -30,7 +30,7 @@ void main() {
 
     testWidgets('renders material style (Android)', (tester) async {
       debugDefaultTargetPlatformOverride = TargetPlatform.android;
-      final explained = await _explain(tester);
+      final explained = await progressIndicatorExplain(tester);
       expect(explained, contains('CircularProgressIndicator'));
       expect(explained, isNot(contains('CupertinoActivityIndicator')));
       debugDefaultTargetPlatformOverride = null;
@@ -38,7 +38,7 @@ void main() {
 
     testWidgets('renders cupertino style (iOS)', (tester) async {
       debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
-      final explained = await _explain(tester);
+      final explained = await progressIndicatorExplain(tester);
       expect(explained, contains('CupertinoActivityIndicator'));
       debugDefaultTargetPlatformOverride = null;
     });
@@ -46,28 +46,28 @@ void main() {
 
   group('useExplainer: false', () {
     const html = '<video><source src="$src"></video>';
-    Future<String> _explain(WidgetTester tester) async {
+    Future<String> explainWithoutExplainer(WidgetTester tester) async {
       final explained = await explain(tester, html, useExplainer: false);
       return explained;
     }
 
     testWidgets('renders video player (Android)', (tester) async {
       debugDefaultTargetPlatformOverride = TargetPlatform.android;
-      final explained = await _explain(tester);
+      final explained = await explainWithoutExplainer(tester);
       expect(explained, contains('Chewie(state: ChewieState)'));
       debugDefaultTargetPlatformOverride = null;
     });
 
     testWidgets('renders video player (iOS)', (tester) async {
       debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
-      final explained = await _explain(tester);
+      final explained = await explainWithoutExplainer(tester);
       expect(explained, contains('Chewie(state: ChewieState)'));
       debugDefaultTargetPlatformOverride = null;
     });
 
     testWidgets('skips video player (linux)', (tester) async {
       debugDefaultTargetPlatformOverride = TargetPlatform.linux;
-      final explained = await _explain(tester);
+      final explained = await explainWithoutExplainer(tester);
       expect(explained, isNot(contains('Chewie(state: ChewieState)')));
       debugDefaultTargetPlatformOverride = null;
     });
